Moved SDK code into rollbar.sdk and updated all import references
