+++ rollbar/sdk/contrib/pyramid/__init__.py
