+++ rollbar/sdk/lib/events.py
