+++ rollbar/sdk/lib/filters/basic.py
