+++ rollbar/sdk/lib/transport.py
