+++ rollbar/sdk/test/test_custom_filters.py
@@ -1,8 +1,8 @@
 import re
 
-from rollbar.lib import events, filters
+from rollbar.sdk.lib import events, filters
 
-from rollbar.test import BaseTest
+from rollbar.sdk.test import BaseTest
 
 
 class CustomFiltersTest(BaseTest):
