+++ rollbar/sdk/test/test_serializable_transform.py
@@ -10,10 +10,10 @@ except ImportError:
     # Python 2.7
     from collections import Mapping
 
-from rollbar.lib import transforms, python_major_version
-from rollbar.lib.transforms.serializable import SerializableTransform
+from rollbar.sdk.lib import transforms, python_major_version
+from rollbar.sdk.lib.transforms.serializable import SerializableTransform
 
-from rollbar.test import BaseTest, SNOWMAN, SNOWMAN_UNICODE
+from rollbar.sdk.test import BaseTest, SNOWMAN, SNOWMAN_UNICODE
 
 if python_major_version() >= 3:
     SNOWMAN = SNOWMAN_UNICODE
